Add a SecurityTrails subdomains discovery task!
