--- lib/tasks/search_security_trails.rb
@@ -1,35 +1,0 @@
-module Intrigue
-module Task
-class SearchSecurityTrails < BaseTask
-
-  def self.metadata
-    {
-      :name => "search_security_trails",
-      :pretty_name => "Search Security Trails",
-      :authors => ["jcran"],
-      :description => "This task hits the Security Trails API and finds matches.",
-      :references => [],
-      :type => "discovery",
-      :passive => true,
-      :allowed_types => ["EmailAddress", "DnsRecord"],
-      :example_entities => [{"type" => "Host", "details" => {"name" => "intrigue.io"}}],
-      :allowed_options => [],
-      :created_types => ["DnsRecord","Info"]
-    }
-  end
-
-  ## Default method, subclasses must override this
-  def run
-    super
-
-    # Make sure the key is set
-    api_key = _get_task_config "security_trails_api_key"
-    entity_name = _get_entity_name
-
-    raise "Not yet implemented"
-
-  end # end run()
-
-end
-end
-end
